+++ docs/changelog.rst
@@ -10,6 +10,9 @@ License
 -----
 :Released: under development
 
+* Improvement: Providing :ref:`needs_external_needs` to allow usage and referencing of external needs.
+  (`#137 <https://github.com/useblocks/sphinxcontrib-needs/issues/137>`_)
+* Bugfix: :ref:`needtable` shows attributes with value ``False`` again.
 * Bugfix: ``:hide:`` and ``:collapse: True`` are working inside :ref:`needimport`.
   (`#284 <https://github.com/useblocks/sphinxcontrib-needs/issues/284>`_,
   `#294 <https://github.com/useblocks/sphinxcontrib-needs/issues/294>`_)